+++ lib/overcommit/hook/post_checkout/base.rb
@@ -7,5 +7,14 @@ module Overcommit::Hook::PostCheckout
 
     def_delegators :@context,
                    :previous_head, :new_head, :branch_checkout?, :file_checkout?
+
+    def skip_file_checkout?
+      @config['skip_file_checkout'] != false
+    end
+
+    def enabled?
+      return false if file_checkout? && skip_file_checkout?
+      super
+    end
   end
 end